UI Revamp, Additional Previews, Consistency in Toasters, Reordering of Content and Golden Ratio Preview Feature
Reorganized the Font and Color sections, including new subtabs for better structure and user experience. Moved the "Why this pairing works" to the right side on desktop. Renamed the "Accessibility" section to "Contrast Check" to be more descriptive of what it actually is. Reorganized info boxes to be visible only when needed for better UI. Added 2 additional previews, going from 7 to now 10 and improved current previews to look more like their intended previews, with improved font size handling in the preview areas. Added the ability to "Undo" more actions, such as when resetting the text to default resetting the font sizes. Lastly, a new Golden Ratio Preview feature was added, allowing the header and body sizes to be previewed with the golden ratio (1.618).